+++ fingerprint-service/obj/Debug/net6.0-windows/FingerprintService.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("CueMiiFingerprintService")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+e537f95211dbb2b0891b02b57e2b7cacbbfc166c")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+997293fe3718394606bc92e249b51c68e8d85352")]
 [assembly: System.Reflection.AssemblyProductAttribute("CueMiiFingerprintService")]
 [assembly: System.Reflection.AssemblyTitleAttribute("CueMiiFingerprintService")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]